Task#4#5#6
Completed the remaining tasks.

For task#4 implemented a dictionary and pagination on file "database.py" and "crud.py"

For task#5 modified the "update_item_by_id" method to not modified anything if it detects more than duplicate of the name.

For task#6 updated the test file "test_items.py" to first create the item.